FRI-46 Project and task acceptance criteria maintenance.

+++ src/main/java/org/snomed/aag/data/services/CriteriaItemService.java
@@ -1,7 +1,10 @@
 package org.snomed.aag.data.services;
 
+import org.snomed.aag.data.Constants;
+import org.snomed.aag.data.domain.CriteriaItem;
+import org.snomed.aag.data.domain.ProjectAcceptanceCriteria;
 import org.snomed.aag.data.repositories.CriteriaItemRepository;
-import org.snomed.aag.domain.CriteriaItem;
+import org.snomed.aag.data.repositories.ProjectAcceptanceCriteriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -16,6 +19,9 @@ public class CriteriaItemService {
 
 	@Autowired
 	private CriteriaItemRepository repository;
+
+	@Autowired
+	private ProjectAcceptanceCriteriaRepository acceptanceCriteriaRepository;
 
 	public void create(CriteriaItem criteriaItem) {
 		repository.save(criteriaItem);
@@ -38,6 +44,11 @@ public class CriteriaItemService {
 	}
 
 	public void delete(CriteriaItem item) {
+		final Page<ProjectAcceptanceCriteria> found = acceptanceCriteriaRepository.findAllBySelectedProjectCriteriaIdsOrSelectedTaskCriteriaIds(item.getId(), item.getId(), Constants.PAGE_OF_ONE);
+		if (!found.isEmpty()) {
+			throw new IllegalArgumentException(format("Criteria can not be deleted, it is used in %s project criteria.", found.getTotalElements()));
+		}
+
 		repository.delete(item);
 	}
 }
